% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{add_inset_map}
\alias{add_inset_map}
\title{Inset map}
\usage{
add_inset_map(
  inset,
  where = "bottomright",
  shape = "rectangular",
  width = 0.22,
  height = 0.25
)
}
\arguments{
\item{inset}{A sf object of the map to plot as thumbnail.}

\item{where}{A character string specifying `where` the inset should be plotted.}

\item{width}{A numeric value specifying the width of the map.}

\item{heigth}{A numeric value specifying the height of the map.}
}
\description{
Add an inset map to an existing plot
}
