% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_functions.R
\name{CJ_par_dt}
\alias{CJ_par_dt}
\title{Parallelized version of the cross-join function from optiRum package}
\usage{
CJ_par_dt(X, Y, nThreads = getOption("mc.cores"))
}
\description{
Parallelized version of the cross-join function from optiRum package
}
