% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{decimalnumcount}
\alias{decimalnumcount}
\title{Count the number of decimal
Modified from https://stat.ethz.ch/pipermail/r-help/2012-July/317676.html}
\usage{
decimalnumcount(x)
}
\arguments{
\item{x}{A character string or numeric vector.}
}
\description{
Count the number of decimal
Modified from https://stat.ethz.ch/pipermail/r-help/2012-July/317676.html
}
