% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{\%dopar\%}
\alias{\%dopar\%}
\title{Parallel operator}
\seealso{
\code{foreach::\link[foreach:do]{\%dopar\%}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{foreach}{\code{\link[foreach:foreach]{\%dopar\%}}}
}}

