% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_evaluation.R
\name{get_best_maxent_model}
\alias{get_best_maxent_model}
\title{utility function: selects the best model based on a set of evaluation metrics.}
\usage{
get_best_maxent_model(model_output, eval_metrics, tolerance = 3)
}
\arguments{
\item{model_output}{A dataframe object returned by the function \code{block_cv_maxent}.}

\item{eval_metrics}{A vector of evaluation metric names. Evaluation metrics available are: `auc`, `omission_rate`,`tss`, `ic`.}

\item{tolerance}{A numeric integer specifying the number of decimals to account for when comparing performance metrics. Default is 3 digits.}
}
\description{
utility function: selects the best model based on a set of evaluation metrics.
}
