% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent_utilities.R
\name{get_maxent_output}
\alias{get_maxent_output}
\title{From an output directory}
\usage{
get_maxent_output(
  maxent_outputdir,
  eval_metrics = c("auc", "omission_rate", "tss"),
  index = 1
)
}
\arguments{
\item{maxent_outputdir}{A character string specifying the path to the output directory where the maxent results are stored.}

\item{eval_metrics}{A character string of the names of the evaluation metrics to retrieve from the maxent results.}
}
\description{
Reads a named list and returns an executing command line for MaxEnt
}
