% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{getCpuUsageRprocs}
\alias{getCpuUsageRprocs}
\title{Get R processes CPU usage}
\usage{
getCpuUsageRprocs(
  targets = c("Rscript.exe", "rsession.exe", "rstudio.exe", "Rgui.exe", "R")
)
}
\description{
Uses 'tasklist' to count the number of `Rscript.exe` instances
}
\details{
Modified from https://stackoverflow.com/questions/47318401/r-how-to-check-how-many-cores-cpu-usage-available
}
