% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{getFreeMemory}
\alias{getFreeMemory}
\title{Get the free available memory}
\usage{
getFreeMemory(units = "KB")
}
\description{
Extract the amount of memory available on the `Windows` and `Linux` machines
}
\details{
Slightly modified from https://stackoverflow.com/questions/27788968/how-would-one-check-the-system-memory-available-using-r-on-a-windows-machine
}
