% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{getNumRprocess}
\alias{getNumRprocess}
\title{Get the number of R processes running}
\usage{
getNumRprocess()
}
\description{
Uses 'tasklist' to count the number of `Rscript.exe` instances
}
\details{
Modified from https://stackoverflow.com/questions/15935931/detect-number-of-running-r-instances-in-windows-within-r
}
