% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_grid.R
\name{grid_pixels}
\alias{grid_pixels}
\title{Build a regular grid from a set of point
modified from https://stackoverflow.com/questions/25547826/generate-regularly-spaced-points-in-polygon}
\usage{
grid_pixels(loc_dat, ext = NULL, dst = 20000)
}
\arguments{
\item{loc_dat}{A two-column matrix,a data.frame, a data.table with longitude and latitude coordinates.}

\item{ext}{A bounding box object specifying the extent of the grid to build. Default is the bounding box of the location data.}

\item{dst}{A numeric value specifying the distance in meters between each grid points.Default is 20000m, i.e. 20km.}
}
\description{
Build a regular grid from a set of point
modified from https://stackoverflow.com/questions/25547826/generate-regularly-spaced-points-in-polygon
}
