% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_utils.R
\name{LL2UTMZone}
\alias{LL2UTMZone}
\title{From Lon/Lat to UTM zone}
\usage{
LL2UTMZone(longitude, latitude, add_latitude_band = FALSE)
}
\arguments{
\item{longitude}{A numeric vector of longitude coordinates in decimal degrees}

\item{latitude}{A numeric vector of latitude coordinates in decimal degrees}
}
\value{
A numeric vector specifying the UTM zone the coordinates belong to.
}
\description{
Converts coordinates from longitude/latitude to UTM zone
}
