% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locs2sf.R
\name{locs2sf}
\alias{locs2sf}
\title{Convert matrix or data frame with point coordinates to a spatial (sf) object}
\usage{
locs2sf(locs, crs = 4326, lon.col = NULL, lat.col = NULL)
}
\arguments{
\item{locs}{A matrix or data frame containing point coordinates data.
If a matrix, the first two columns will be assumed to contain longitude
and latitude coordinates, respectively.
If a data frame, the function will try to guess the columns containing the coordinates
based on column names, unless \code{lon.col} and \code{lat.col} are provided.}

\item{crs}{A character string, number (EPSG value), or \code{crs} object specifying the
coordinate reference system (see \code{\link[sf:st_crs]{sf::st_crs()}} or \url{https://spatialreference.org}).
Default is geographic (unprojected) coordinates, datum WGS84 (EPSG = 4326).}

\item{lon.col}{Character (optional). Name of the column containing longitude data.}

\item{lat.col}{Character (optional). Name of the column containing latitude data.}
}
\value{
An \code{\link[sf:sf]{sf::sf()}} object.
}
\description{
Convert matrix or data frame with point coordinates to a spatial (sf) object
}
\examples{
locs <- matrix(runif(20), ncol = 2)
locs2sf(locs)

locs <- data.frame(species = rep("Laurus nobilis", 10), x = runif(10), y = runif(10))
locs2sf(locs)
}
