% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locs2sp.R
\name{locs2sp}
\alias{locs2sp}
\title{Convert matrix or dataframe with coordinates to a spatial object}
\usage{
locs2sp(locs, proj = "+init=epsg:4326", lon.col = NULL, lat.col = NULL)
}
\arguments{
\item{locs}{A matrix or dataframe containing coordinates data. If a matrix, the first two columns will be assumed to contain longitude and latitude coordinates, respectively. If a dataframe, the function will try to guess the columns containing the coordinates based on column names.}

\item{proj}{Character string specifying the projection (see \code{\link[sp]{proj4string}} or
\url{http://spatialreference.org}). Default is geographic (unprojected) coordinates, datum WGS84.}

\item{lon.col}{Character (optional). Name of the column containing longitude data.}

\item{lat.col}{Character (optional). Name of the column containing latitude data.}
}
\value{
A SpatialPoints or SpatialPointsDataFrame.
}
\description{
Convert matrix or dataframe with coordinates to a spatial object
}
\examples{
locs <- matrix(runif(20), ncol = 2)
locs.sp <- locs2sp(locs)

locs <- data.frame(species = rep("Laurus nobilis", 10), x = runif(10), y = runif(10))
locs.sp <- locs2sp(locs)
}

