% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locs2vect.R
\name{locs2vect}
\alias{locs2vect}
\title{Convert matrix or data frame with point coordinates to a SpatVector object}
\usage{
locs2vect(locs, crs = 4326, lon.col = NULL, lat.col = NULL)
}
\arguments{
\item{locs}{A matrix or data frame containing point coordinates data.
If a matrix, the first two columns will be assumed to contain longitude
and latitude coordinates, respectively.
If a data frame, the function will try to guess the columns containing the coordinates
based on column names, unless \code{lon.col} and \code{lat.col} are provided.}

\item{crs}{A number specifying the EPSG code (see \url{https://spatialreference.org}).
Default is geographic (unprojected) coordinates, datum WGS84 (EPSG = 4326).}

\item{lon.col}{Character (optional). Name of the column containing longitude data.}

\item{lat.col}{Character (optional). Name of the column containing latitude data.}
}
\value{
A \code{\link[terra:SpatVector-class]{terra::SpatVector()}} object.
}
\description{
Convert matrix or data frame with point coordinates to a SpatVector object
}
\examples{
locs <- matrix(runif(20), ncol = 2)
locs2vect(locs)

locs <- data.frame(species = rep("Laurus nobilis", 10), x = runif(10), y = runif(10))
locs2vect(locs)
}
