% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_geographic_background.R
\name{make_projection_domain}
\alias{make_projection_domain}
\title{Estimating the projection area of a species}
\usage{
make_projection_domain(
  bg_dat,
  output_name = NULL,
  output_dir = NULL,
  dissolve = FALSE,
  extent_only = FALSE,
  use_ecoregions = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{bg_dat}{A spatial object. It can be a raster or a sf surface object (i.e. polygon).}

\item{output_name}{An optional character string specifying the name to be given to the domain.}

\item{output_dir}{An optional directory where to save the domain defined.}

\item{dissolve}{A logical. Should the borders between adjacent polygons be dissolved ? Default is \code{FALSE}.}

\item{use_ecoregions}{A logical. Should ecoregions be used instead of biomes ? Default is TRUE (i.e. ecoregions are used by default).}
}
\value{
An sf object
}
\description{
Extract all ecoregions falling inside the bounding box of a raster or polygon
}
