% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{map_lfd}
\alias{map_lfd}
\title{utility function: map function that transforms input by applying a function to each element and returning
a data frame with a number of columns equal to the number of element in the input.}
\usage{
map_lfd(x, fun, ...)
}
\description{
utility function: map function that transforms input by applying a function to each element and returning
a data frame with a number of columns equal to the number of element in the input.
}
