% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{map_raster}
\alias{map_raster}
\title{Map the species distribution}
\usage{
map_raster(
  x,
  y = NULL,
  inset = NULL,
  outputdir,
  main_title = NULL,
  legend_title = NULL,
  output_format = "pdf",
  inset_shape = "rectangular",
  logz = FALSE,
  reproject = FALSE,
  projection = "+proj=wintri",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A RasterLayer or stars object, or a path to the raster layer file.}

\item{y}{An optional sf object describing spatial features to be displayed on the map: points, polygons,...etc.}

\item{inset}{An optional inset map to be added to the plot. Default is `NULL`.}

\item{outputdir}{A character string specifying the output directory where the map created will be stored.}

\item{main_title}{A character string specifying the title of the map.}

\item{legend_title}{A character string specifying the title of the legend.}

\item{output_format}{A character string specifying in which format the map should be exported. Default is `pdf` format.}

\item{inset_shape}{A character string specifying the shape of the thumbnail for the inset map: `rectangular` (default) or `circular`}

\item{logz}{A logical. Should the raster values be log-transformed ? Default is `FALSE`.}

\item{reproject}{A logical. Should the raster layer be reprojected. Ignored if argument `projection` and the raster projection are the same.}

\item{projection}{A character string specifying the new projection in which the raster must be reprojected. Ignored if \code{reproject=FALSE}.}

\item{...}{Additional graphic parameters}
}
\description{
Plot the raster map of the probability of occurrence of a species
}
