% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_utils.R
\name{maskCover}
\alias{maskCover}
\title{Crop/mask raster with a spatial object while accounting for cell coverage}
\usage{
maskCover(x, y, do.crop = TRUE, ...)
}
\arguments{
\item{x, }{A `Raster*` object}

\item{y, }{A `Spatial*` object}
}
\value{
A `Raster*` object
}
\description{
crop then mask a raster object with a spatial shape while making sure that boundary cells in contact with the shape are not masked out
}
