% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_utils.R
\name{meters2degrees}
\alias{meters2degrees}
\title{From meters to decimal degrees}
\usage{
meters2degrees(x, dist, ...)
}
\arguments{
\item{x}{An sf object representative of a location from where the distance in degree decimal will be calculated. It can be points, polygons,...etc}

\item{dist}{A numeric value specifying the distance in meters to be converted in degrees}
}
\description{
Converts distance in meters to distance in decimal degrees given a geometry
}
