% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{min.col}
\alias{min.col}
\title{utility function: find the minimum position for each row of a matrix, breaking ties at random.}
\usage{
\method{min}{col}(m, ...)
}
\description{
utility function: find the minimum position for each row of a matrix, breaking ties at random.
}
