% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{modal}
\alias{modal}
\title{Most frequent value(s)}
\usage{
modal(x, n = 1, ...)
}
\arguments{
\item{x}{A vector}

\item{n}{A numeric integer specifying the number of values to return. Default is 1, i.e. the most frequent value.}

\item{...}{Additional parameters to be passed to the function \code{table}.}
}
\description{
Find the most frequent n value(s) from a vector
}
\seealso{
\code{table}
}
