% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{\%>\%}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\title{Pipe operator(s)}
\usage{
lhs \%>\% rhs
}
\seealso{
\code{magrittr::\link[magrittr:pipe]{\%>\%}}.

\code{magrittr::\link[magrittr:pipe]{\%<>\%}}

\code{magrittr::\link[magrittr:pipe]{\%$\%}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:exposition]{\%$\%}}, \code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}
}}

