% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gmap.R
\name{plot_gmap}
\alias{plot_gmap}
\title{Plot Google Earth images obtained through dismo::gmap}
\usage{
plot_gmap(x, maxpixels = 5e+05, ext = NULL, interpolate = FALSE, axes,
  main, xlab = "", ylab = "", asp, add = FALSE, addfun = NULL,
  zlim = NULL, zlimcol = NULL, ...)
}
\arguments{
\item{x}{RasterLayer, as obtained through \code{\link[dismo]{gmap}}.}

\item{maxpixels}{integer > 0. Maximum number of cells to use for the plot. If \code{maxpixels < ncell(x)}, \code{sampleRegular} is used before plotting. If \code{gridded=TRUE} maxpixels may be ignored to get a larger sample.}

\item{ext}{An extent object to zoom in a region}

\item{interpolate}{Logical. Should the image be interpolated (smoothed)?}

\item{axes}{not used}

\item{main}{character. Main plot title}

\item{xlab}{Optional. x-axis label}

\item{ylab}{Optional. y-axis label}

\item{asp}{not used}

\item{add}{Logical. Add to current plot?}

\item{addfun}{Function to add additional items such as points or polygons to the plot (map). Typically containing statements like "points(xy); plot(polygons, add=TRUE)". This is particularly useful to add something to each map when plotting a multi-layer Raster* object.}

\item{zlim}{not used}

\item{zlimcol}{not used}

\item{...}{Graphical parameters. Any argument that can be passed to \code{\link[graphics]{rasterImage}}.}
}
\value{
A plot.
}
\description{
This function is very slightly modified from \code{.plotCT} function in \code{raster} package to avoid producing an empty plot before the actual Google image.
}
\examples{
library(dismo)
g <- gmap("France")
plot_gmap(g)
}
