% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_grid.R
\name{pointsTogrid}
\alias{pointsTogrid}
\title{Convert a set of point to a grid
adapted from https://stackoverflow.com/questions/25547826/generate-regularly-spaced-points-in-polygon}
\usage{
pointsTogrid(loc_dat, domain, dst = 20000)
}
\arguments{
\item{loc_dat}{A two-column matrix,a data.frame, a data.table with longitude and latitude coordinates.}

\item{domain}{A sf object specifying the extent of the grid to build. Default is the bounding box of the location data.}

\item{dst}{A numeric value specifying the distance in meters between each grid points.Default is 20000m, i.e. 20km.}
}
\description{
Convert a set of point to a grid
adapted from https://stackoverflow.com/questions/25547826/generate-regularly-spaced-points-in-polygon
}
