% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prediction.R
\name{project_geoModel}
\alias{project_geoModel}
\title{Project a fitted geographic distance-based model}
\usage{
project_geoModel(
  domain,
  model,
  output_directory,
  project_res = NULL,
  output_name = NULL,
  fun = function(x, na.rm = TRUE) 1 - prod(1 - x, na.rm = na.rm),
  save_output = TRUE
)
}
\arguments{
\item{domain}{An object of class `RasterLayer` representing the spatial grid domain where to interpolate the model}

\item{model}{An object of class `GeographicDistance` or `InvDistWeightModel`}

\item{output_directory}{An optional character string specifying the directory where to save the raster of interpolated values}

\item{project_res}{A numeric value specifying the resolution of the projected raster in decimal degree. Default is the same as raster layer `domain`.}

\item{save_output}{A logical. Should the output raster be saved ? Ignored if `output_directory` is not provided. Default is `TRUE`.}

\item{return}{An object of class `RasterLayer` with values interpolated by the model}
}
\description{
This function is a wrapper for the function `predict` from the `raster` package
}
