% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_utils.R
\name{read_layers}
\alias{read_layers}
\title{utility function: read multiple raster layers from a directory.}
\usage{
read_layers(path, nlayers = NULL)
}
\arguments{
\item{path}{A character string specifying the path to the directory where the raster layers are stored}

\item{nlayers}{A numeric integer specifying the number of layers to read. Default is all layers.}
}
\description{
utility function: read multiple raster layers from a directory.
}
