% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSDM-package.R
\docType{package}
\name{rSDM-package}
\alias{rSDM}
\alias{rSDM-package}
\title{rSDM: Species distribution and niche modelling in R}
\description{
Functions for niche modelling and SDM.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pakillo.github.io/rSDM/}
  \item Report bugs at \url{https://github.com/Pakillo/rSDM/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com}

}
\keyword{internal}
