% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{run_maxentModel}
\alias{run_maxentModel}
\title{Run maxent model workflow}
\usage{
run_maxentModel(
  loc_dat,
  species_name = NULL,
  outputdir,
  newdata = NULL,
  verbose = TRUE,
  maxent_settings = list(),
  do.map = FALSE,
  ...
)
}
\arguments{
\item{loc_dat}{A character string specifying the path to a csv file with longitude and latitude coordinates of occurrence records.}

\item{outputdir}{A character string specifying the directory where to save the raster of interpolated values.}

\item{newdata}{A character string specifying the directory where the new environmental raster layers for projection are stored.
Default is `NULL` i.e. projection will use environmental layers provided for calibration.}

\item{...}{Additional parameters to be passed to \code{block_cv_maxent} or \code{make_geographic_block}}

\item{algorithm}{A character string specifying the name of the algorithm to use.}
}
\value{
An integer number. `0` in case of success otherwise return `-1`
}
\description{
Uses masked geographically structured cross-validation method for evaluating and selecting a maxent model
}
