% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{run_pointModel}
\alias{run_pointModel}
\title{Run point-based model}
\usage{
run_pointModel(
  loc_dat,
  species_name = NULL,
  outputdir,
  project_res = 0.1666667,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{loc_dat}{A character string specifying the path to a csv file with longitude and latitude coordinates of occurrence records.}

\item{outputdir}{A character string specifying the directory where to save the output raster.}

\item{project_res}{A numeric value specifying the resolution of the output raster in decimal degree. default is 0.1666667 i.e ~ 20km}
}
\value{
An integer number. `0` in case of success otherwise `-1` in case of failure.
}
\description{
Uses known occurrence locations as its own.
}
