% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_SDMWorkflow.R
\name{run_SDMWorkflow}
\alias{run_SDMWorkflow}
\title{Running SDM Workflow}
\usage{
run_SDMWorkflow(
  base_directory,
  species_directory,
  env_directory,
  force_sorting = FALSE,
  workflow = system.file("extdata", "workflows", "1UsefulPlants_Workflow.R", package =
    "UsefulPlants"),
  run_name = "Test",
  mc_cores = NULL
)
}
\arguments{
\item{base_directory}{A character string specifying the path to the directory where to store the outputs of the workflow}

\item{species_directory}{A character string specifying  the path to the directory where are stored the species occurrence records.}

\item{env_directory}{A character string specifying the path to the directory where are stored the environmental layers.}

\item{force_sorting}{A logical. Should the sorting of occurrence records be repeated ? Only useful if `species_directory` contains new or updated occurrence data. Default is `FALSE`.}

\item{workflow}{A character string specifying the path to the modelling script to run}
}
\value{
None
}
\description{
Execute a modelling pipeline R script
}
\examples{
run_SDMWorkflow()
}
