% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sea2land.R
\name{sea2land}
\alias{sea2land}
\title{Move occurrences to the nearest raster cell with data}
\usage{
sea2land(locs, ras, showmap = TRUE)
}
\arguments{
\item{locs}{A matrix, dataframe or SpatialPointsDataFrame containing coordinates of species occurrence. #' If a simple (non-spatial) dataframe, coordinates must be contained in columns named as ("x", "y") or #' #' ("lon", "lat"). If a matrix, coordinates must be the first two columns (longitude, latitude).}

\item{ras}{\code{Raster*} object containing the environmental (climatic) layers.}

\item{showmap}{Logical. Show map with original and new coordinates?}
}
\value{
A SpatialPointsDataFrame with corrected coordinates.
}
\description{
Move point occurrences falling 'on sea' (i.e. NA) to the nearest raster cell with data ('land').
}
\examples{
\\donotrun{
data(acaule)
ras <- getData('worldclim', var = 'tmin', res = 10)
check.coords <- sea2land(locs = acaule, ras)
}
}
\author{
F. Rodriguez-Sanchez.
}
\seealso{
\url{https://github.com/SEEG-Oxford/seegSDM/blob/master/man/nearestLand.Rd} and
\url{http://stackoverflow.com/questions/26652629/extracting-a-value-from-a-raster-for-a-specific-point-based-on-the-closest-cell/26688361#26688361}.
}

