% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_functions.R
\name{st_par_covered_by}
\alias{st_par_covered_by}
\title{Parallelized version of geometric binary predicates function st_covered_by}
\usage{
st_par_covered_by(x, y, nchunks = 10, ncores = getOption("mc.cores"))
}
\description{
Parallelized version of geometric binary predicates function st_covered_by
}
