% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_functions.R
\name{st_par_intersection}
\alias{st_par_intersection}
\title{Compute geometric binary predicates of intersection between 2 simple features geometry sets in parallel
then return the intersecting features}
\usage{
st_par_intersection(x, y, nchunks = 10, ncores = getOption("mc.cores"))
}
\description{
Compute geometric binary predicates of intersection between 2 simple features geometry sets in parallel
then return the intersecting features
}
