% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_functions.R
\name{st_par_intersects}
\alias{st_par_intersects}
\title{Parallel version of Geometric binary predicates from the package `sf`}
\usage{
st_par_intersects(x, y, nchunks = 10, ncores = getOption("mc.cores"))
}
\arguments{
\item{x}{An object of class `sf`, `sfc` or `sfg`}

\item{y}{An object of class `sf`, `sfc` or `sfg`}

\item{nchunks}{A numeric value specifying the number of subgeometries the x object should be split into.}

\item{ncores}{A numeric integer specifying the number of cpu cores to use for parallel processing.}
}
\value{
A sparse list, with list element i an integer vector with all indices j for which intersection(x[i],y[i]) is `TRUE` if they intersect, hence integer(0) if none of them is `TRUE`.
}
\description{
Compute geometric binary predicates of intersection between 2 simple features geometry sets in parallel
}
