% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_functions.R
\name{st_par_within}
\alias{st_par_within}
\title{Parallelized version of geometric binary predicates function st_within}
\usage{
st_par_within(x, y, nchunks = 10, ncores = getOption("mc.cores"))
}
\description{
Parallelized version of geometric binary predicates function st_within
}
