% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_geographic_background.R
\name{st_rectangle}
\alias{st_rectangle}
\title{utility function that creates a rectangle (polygon) from a bounding box}
\usage{
st_rectangle(bbx, tolerance = 0.1)
}
\arguments{
\item{bbx}{A bbox object}

\item{tolerance}{A numeric number specifying the buffer distance in decimal degree allowed around the rectangle}
}
\description{
utility function that creates a rectangle (polygon) from a bounding box
}
