% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdm_utils.R
\name{st_remove_holes}
\alias{st_remove_holes}
\title{Remove hole(s)
from https://stackoverflow.com/questions/52654701/removing-holes-from-polygons-in-r-sf
Remove holes from the geometry of a polygon}
\usage{
st_remove_holes(x)
}
\arguments{
\item{x}{A sf or sfc object.}
}
\description{
Remove hole(s)
from https://stackoverflow.com/questions/52654701/removing-holes-from-polygons-in-r-sf
Remove holes from the geometry of a polygon
}
