% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_evaluation.R
\name{train_geoModel}
\alias{train_geoModel}
\title{Train geographic model}
\usage{
train_geoModel(
  loc_dat,
  bg_dat = NULL,
  algorithm = "idw",
  coordHeaders = NULL,
  outputdir,
  species_name = NULL,
  grid_res = 0.01666667,
  save_bg = TRUE,
  ...
)
}
\arguments{
\item{loc_dat}{A two-column matrix,a data.frame, a data.table or a csv file with longitude and latitude coordinates of occurrence records.}

\item{algorithm}{A character string specifying the name of the algorithm to use. Default is `idw` i.e. Inverse-Distance Weighted.}

\item{coordHeaders}{[Optional] A character string vector of length two giving the names of the coordinates in \code{loc_dat}}

\item{outputdir}{A character string specifying the directory where to save the outputs of the model.}

\item{species_name}{[Optional] A character string specifying the name of the species being modelled.}

\item{grid_res}{A numeric value specifying the resolution of the output raster layer in decimal degree. Default is 0.01666667 (~ 2 km).}

\item{save_bg}{A logical. Should the background be saved ? Default is `TRUE`.}
}
\description{
Calibrate and evaluate a geographic model using interpoint distance-based algorithm
}
