% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_evaluation.R
\name{train_Maxent}
\alias{train_Maxent}
\title{Train maxent model}
\usage{
train_Maxent(
  loc_dat,
  env_dat,
  outputdir,
  coordHeaders = NULL,
  species_name = NULL,
  save_bg = TRUE,
  ...
)
}
\arguments{
\item{loc_dat}{A two-column matrix,a data.frame, a data.table or a csv file with longitude and latitude coordinates of occurrence records.}

\item{env_dat}{A RasterStack object or a character string specifying the path to a directory with environmental raster layers.}

\item{outputdir}{A character string specifying the directory where to save the outputs of the model.}

\item{coordHeaders}{[Optional] A character string vector of length two giving the names of the coordinates in \code{loc_dat}}

\item{species_name}{[Optional] A character string specifying the name of the species being modelled.}

\item{save_bg}{A logical. Should the background be saved ? Default is `TRUE`.}

\item{...}{Additional parameters to be passed to \code{block_cv_maxent} or \code{make_geographic_block}}
}
\description{
Calibrate and Evaluate a maxent model using the masked geographically structured cross-validation method for evaluation
}
\details{
Partition both occurrence records and background into evaluation bins based on some spatial rules.
}
