% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{unregister}
\alias{unregister}
\title{Unregister a parallel backend}
\usage{
unregister()
}
\description{
Provides a mechanism to unregister a foreach backend
}
\details{
https://stackoverflow.com/questions/25097729/un-register-a-doparallel-cluster
}
