% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/occmap.R
\name{occmap}
\alias{occmap}
\title{Plot map of species occurrences}
\usage{
occmap(locs, bg = "google", projection, pcol = "red", alpha = 1,
  psize = 1, add = FALSE, mapmisc_server = "maptoolkit", ...)
}
\arguments{
\item{locs}{A matrix, dataframe or SpatialPointsDataFrame containing coordinates of species occurrences. If a simple (non-spatial) dataframe,
coordinates must be contained in columns named as ("x", "y") or ("lon", "lat"). If a matrix, coordinates must be the first two columns (longitude, latitude).}

\item{bg}{Character. Either
'google' for Google maps background (using dismo::gmap),
'coast' for coastlines (coastsCoarse shapefile from rworldmap package),
'ggmap' for any of the maps provided by get_map function in ggmap package,
'leaflet' for an interactive HTML map using leaflet, or
'mapmisc' for using any of the layers available in \code{mapmisc} package, or
'kml' for producing a KML file to be opened with Google Earth.
Alternatively, a Raster object to be used as background.}

\item{projection}{CRS object giving the projection system of species occurrences.
Leave missing if species occurrences are in geographical projection, or 'locs' is a SpatialPoints* object containing projection information (proj4string).}

\item{pcol}{Colour to be used for points representing species' occurrences. Default is "red".}

\item{alpha}{Colour transparency for points between 0 (fully transparent) and 1 (fully opaque).}

\item{psize}{Point size for depicting species' occurrences. Default is 1 (cex=1).}

\item{add}{Logical. Add these occurrences to previous map (e.g. for a new species?). Default is FALSE. Note this feature doesn't work for all map types.}

\item{mapmisc_server}{character. Server/type of background map to be used when bg = "mapmisc". Run \code{osmTiles()} to see all the available layers.}

\item{...}{additional parameters to be passed to
dismo::gmap if bg == 'google'
plot if bg == 'coast'
ggmap::get_map if bg == 'ggmap'
leafletR::leaflet if bg == 'leaflet'
raster::plot or raster::plotRGB if bg == 'mapmisc'
See these functions help files for details.}
}
\value{
A map (plot), unless \code{bg = 'KML'} in which case a kmz file is saved.
}
\description{
Plot map of species occurrences on top of different background layers.
}
\details{
If using ggmap and stamen maps, large regions seem to give problems.
}
\examples{
# Using acaule dataset from dismo package:
data(acaule)
occmap(locs=acaule)
occmap(locs=acaule, bg="google")
occmap(locs=acaule, bg="google", type="satellite")   # using options from dismo:gmap


# Watercolor map from Stamen using ggmap
# NB: use small regions, otherwise give error to download map tiles
locs_redux <- subset(acaule, lon>-80 & lon< -60 & lat>-30 & lat< -10)
occmap(locs=locs_redux,
           bg="ggmap", maptype='watercolor', source='stamen',
           pcol="darkgreen", psize=4)


# Plot occurrences in a specific country:
occmap(locs=subset(acaule, country=="Bolivia"))

# Plot occurrences within given coordinates
occmap(locs=subset(acaule, lon>-80 & lon< -60 & lat>-30 & lat< -10))
# can use click() aftewards to identify specific points


# Add transparency to points
occmap(locs=acaule, pcol="red", alpha=0.7)

# Save plot directly to file
pdf("map.pdf", paper="a4r")
occmap(locs=acaule, type="satellite", scale=2)
dev.off()

## Providing spatial objects ##
data(meuse)
coordinates(meuse) <- ~x+y
proj4string(meuse) <- CRS("+init=epsg:28992")
occmap(meuse)

#alternatively, provide projection argument:
data(meuse)
coordinates(meuse) <- ~x+y
occmap(meuse, projection = CRS("+init=epsg:28992"))
}
\author{
F. Rodriguez-Sanchez. Code for using spplot taken from Oscar Perpinan's blog.
}

