% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_in_cell.R
\name{point_in_cell}
\alias{point_in_cell}
\title{Do points (occurrences) fall within raster cells?}
\usage{
point_in_cell(locs, ras, layer = 1)
}
\arguments{
\item{locs}{A matrix, dataframe, SpatialPoints or SpatialPointsDataFrame containing coordinates of species occurrences. If locs is a matrix or dataframe, it will be converted to a spatial object using \code{\link[=locs2sp]{locs2sp()}}.}

\item{ras}{Raster* object}

\item{layer}{Integer. For RasterStack and RasterBrick objects, which layer to use to compare with point locations (default = 1).}
}
\value{
A logical vector.
}
\description{
This function examines which points fall within a raster cell with data (not NA). Returns TRUE for points falling in a raster cell with data, and FALSE otherwise.
}
