% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points2nearestcell.R
\name{points2nearestcell}
\alias{points2nearestcell}
\title{Move occurrences to the nearest raster cell with data}
\usage{
points2nearestcell(locs, ras, layer = 1, move = TRUE, distance = NULL,
  showchanges = TRUE, showmap = TRUE, leaflet = FALSE)
}
\arguments{
\item{locs}{A SpatialPoints or SpatialPointsDataFrame.}

\item{ras}{\code{Raster*} object.}

\item{layer}{Integer. For RasterStack and RasterBrick objects, layer to use.}

\item{move}{Logical. Change coordinates of points to those of the nearest raster cells?}

\item{distance}{Numeric (optional). Maximum distance to move points. Point coordinates are only changed if the distance to the nearest raster cell is below \code{distance}.}

\item{showchanges}{Logical. Print table with old and new coordinates.}

\item{showmap}{Logical. Show map with original and new coordinates?}

\item{leaflet}{Logical. If TRUE, show leaflet map.}
}
\value{
A SpatialPointsDataFrame (with corrected coordinates if move is TRUE).
}
\description{
Move point occurrences falling in raster cells without data (i.e. NA) to the nearest raster cell with data.
}
\examples{
\dontrun{
data(acaule)
locs <- locs2sp(acaule)
ras <- getData('worldclim', var = 'tmin', res = 10)
check.coords <- points2nearestcell(locs, ras)
}
}
\seealso{
\url{https://github.com/SEEG-Oxford/seegSDM/blob/master/man/nearestLand.Rd} and
\url{http://stackoverflow.com/questions/26652629/extracting-a-value-from-a-raster-for-a-specific-point-based-on-the-closest-cell/26688361#26688361}.
}

